Update profileImage.js
If the name of a user is too long, the last letters are replaced with three dots: "..." to avoid this, you could set the font size of the username to 60px and that of the tag to 40px. An example of a name that is too long can be seen in the image below.  
https://i.imgur.com/zQdimDc.png